# backend/docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Club Booking SaaS Platform API
  description: >
    Multi-tenant booking engine for Players and Club Owners.
    
    ## Authentication
    All protected endpoints use JWT stored in an **HttpOnly cookie** named `access_token_cookie`.
    The browser automatically sends the cookie with each request – no manual token management is required.
    Endpoints that require specific roles (e.g., `owner`, `player`) are marked accordingly.

    ## Implementation Deviations Logged
    1. Operating hours fields (`open_time`, `close_time`, `slot_duration_minutes`) are flat fields on `Club` and `Court` models, not nested objects.
    2. Overlap validation for Bookings is performed at the application layer via Python queries instead of PostgreSQL GiST constraints, as SQLite is used.
    3. Additional modules implemented and documented: Memberships, Events, Payments, Attendance, and Notifications.
  version: 2.1.0
servers:
  - url: http://127.0.0.1:5000/api/v1
    description: Local Flask Development Server
security:
  - cookieAuth: []   # JWT stored in HttpOnly cookie
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: JWT authentication cookie (HttpOnly)
  schemas:
...
    OperatingHours:
      type: object
      properties:
        open_time:
          type: string
          example: "06:00"
          description: "HH:MM format"
        close_time:
          type: string
          example: "22:00"
          description: "HH:MM format"
        slot_duration_minutes:
          type: integer
          example: 60
          description: "Granularity of bookable slots"
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        owner_id:
          type: integer
        operating_hours:
          $ref: '#/components/schemas/OperatingHours'
    Court:
      type: object
      properties:
        id:
          type: integer
        club_id:
          type: integer
        name:
          type: string
        is_active:
          type: boolean
        operating_hours_override:
          $ref: '#/components/schemas/OperatingHours'

    # ----- Memberships -----
    MembershipPlan:
      type: object
      properties:
        id:
          type: integer
        club_id:
          type: integer
        name:
          type: string
        price_monthly:
          type: number
          format: float
        benefits:
          type: string
        is_active:
          type: boolean
...
paths:
...
  /clubs:
    get:
      summary: List all clubs (public)
      security: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Optional search by club name
      responses:
        '200':
          description: Array of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
    post:
      summary: Create a new club (Owner Only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address]
              properties:
                name: { type: string }
                address: { type: string }
                open_time: { type: string, example: "06:00" }
                close_time: { type: string, example: "22:00" }
                slot_duration_minutes: { type: integer, example: 60 }
      responses:
        '201':
          description: Club created successfully
        '400':
          description: Validation error
        '409':
          description: Owner already has a club
  /clubs/{clubId}:
    put:
      summary: Update club settings (Owner Only)
      parameters:
        - name: clubId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                address: { type: string }
                open_time: { type: string }
                close_time: { type: string }
                slot_duration_minutes: { type: integer }
      responses:
        '200':
          description: Club updated successfully
        '403':
          description: Not authorized
  /clubs/{clubId}/courts:
    get:
      summary: List active courts for a club (public)
      parameters:
        - name: clubId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Array of active courts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Court'
        '404':
          description: Club not found
  /clubs/my-club:
    get:
      summary: Get owner's own club (Owner Only)
      responses:
        '200':
          description: Club data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '404':
          description: No club found

  # --- Court CRUD (Owner Only) ---
  /clubs/courts:
    get:
      summary: List owner's courts with club info (Owner Only)
      responses:
        '200':
          description: Courts and associated club
          content:
            application/json:
              schema:
                type: object
                properties:
                  club:
                    $ref: '#/components/schemas/Club'
                  courts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Court'
    post:
      summary: Create a court (and optionally a club if none exists)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [court_name]
              properties:
                court_name: { type: string }
                club_name: { type: string, description: "Required if owner has no club yet" }
                club_address: { type: string, description: "Required if owner has no club yet" }
      responses:
        '201':
          description: Court created
        '400':
          description: Validation error or club required but missing
  /clubs/courts/{courtId}:
    put:
      summary: Update a court (Owner Only)
      parameters:
        - name: courtId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [court_name]
              properties:
                court_name: { type: string }
                is_active: { type: boolean, default: true }
                open_time_override: { type: string, nullable: true }
                close_time_override: { type: string, nullable: true }
                slot_duration_override: { type: integer, nullable: true }
      responses:
        '200':
          description: Court updated successfully
        '404':
          description: Court not found or not owned
    delete:
      summary: Delete a court (Owner Only)
      parameters:
        - name: courtId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Court deleted successfully
        '404':
          description: Court not found or not owned

  # ================= AVAILABILITY =================
...
  /memberships/plans:
    get:
      summary: Get membership plans (public)
      parameters:
        - name: club_id
          in: query
          schema: { type: integer }
          description: Filter by club
      responses:
        '200':
          description: List of membership plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MembershipPlan'
  /memberships/subscribe:
    post:
      summary: Subscribe to a membership plan (JWT required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id]
              properties:
                plan_id: { type: integer }
                auto_renew: { type: boolean, default: false }
      responses:
        '201':
          description: Subscribed successfully
        '400':
          description: Validation error
        '404':
          description: Plan not found
        '409':
          description: Already has active membership for this club
...
  /memberships/{membershipId}/cancel:
    post:
      summary: Cancel a membership (JWT required – ownership enforced)
      parameters:
        - name: membershipId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Membership cancelled successfully
        '403':
          description: Not authorized to cancel this membership
        '404':
          description: Membership not found

  # ================= EVENTS =================
...